Released [v3.3.5.6] backup on 210223

+++ DDAHospApi/DataAccess/ITT_OrderNoMapping_Manage.cs
@@ -10,6 +10,6 @@ namespace DDAApi.DataAccess
     {
         Task<int> AddOrderNoMapping(TT_OrderNoMapping mapping);
 
-        TT_OrderNoMapping GetMapping(string orderNo, int ttId);
+        TT_OrderNoMapping GetMapping(string orderNo, long ttId);
     }
 }
