+++ DDAHospApi/HospModel/TT_OrderNoMapping.cs
@@ -11,7 +11,7 @@ namespace DDAApi.HospModel
     {
         public Int64 Id { get; set; }
         public string OrderNo { get; set; }
-        public int TTId { get; set; }
+        public long TTId { get; set; }
         public string PlatOrderNo { get; set; }
         public string PlatName { get; set; }
